Address review feedback: reuse header buffer, cover buffer loops, decouple from #10085
Follow-up to the review on #10297. Three non-blocking findings plus the process
question, all addressed.

Reuse the header line buffer (review: per-frame allocation)
  ReadHeaderLineAsync allocated a fresh 256-byte array per header line, roughly
  three per frame, on the hottest read path -- server mode emits a
  testing/testUpdates/tests notification per test. The buffer is now an instance
  field grown in place. This is safe for the same reason _readBufferOffset and
  _readBufferCount already are: reads are single-threaded by construction,
  driven by exactly one read loop.

Document the unbounded Content-Length as a conscious decision (review:
observation)
  commandSize is still trusted unbounded. That is deliberate: server mode is a
  loopback channel to a test host this process launched, and legitimate bodies
  are unbounded in principle, so a cap would be a guess that risks rejecting
  valid traffic. Recorded in a comment rather than changed, and noted that the
  byte-based read more than halves the previous worst case, which rented
  commandSize chars.

Cover the multi-iteration buffer paths (review: coverage gap)
  The previous tests all used frames under ReadBufferSize and header lines under
  HeaderLineInitialCapacity, so the two loops this change introduced only ever
  ran a single iteration. Three tests added:
    - a body far larger than the read buffer, so ReadExactlyAsync refills
      several times and the body spans buffer boundaries;
    - a header line far longer than the initial header buffer, so the growth
      path runs through several doublings;
    - a peer that announces a body then disconnects mid-way, which must surface
      as ReadAsync returning null rather than hanging or throwing.
  The first and third fail against the pre-fix transport; the second guards a
  path this change introduced.

Decouple the fix from the source-package work (review: process question)
  The src/ half applies to main as-is, but the tests lived in the new
  ServerClient.UnitTests project, coupling a shipping-server bug fix to the
  packaging change. They now live in Microsoft.Testing.Platform.UnitTests,
  merged into the existing TcpMessageHandlerTests. That project targets net462
  alongside net8.0/net9.0, so both the #if NETCOREAPP (System.Text.Json) and
  #else (Jsonite) compilation branches are still covered, and this PR no longer
  depends on #10085 for anything.

Assertions were reworked to compare the JSON-RPC method rather than a params
payload: params binding requires a registered method, so the previous
payload-based assertions were not portable to a transport-level test outside the
ServerClient harness. The non-ASCII content moved into the method name, which
both formatters always bind, so the assertions still compare every multi-byte
character rather than merely checking that a frame arrived.

Copilot-Session: 365edd84-6344-4ce4-9048-2c1ed1b7d939
